CR-2008-11-06-0: Bugfix:     - do not send a "exit\n" when a connection is                                not responding anymore                              - enhanced CHECK_EXPECT_MATCH_MAX_BUFFER size                              - reworked delete_temp_script_files() procedure                              - reworked wait_for_remote_file() procedure                  Review:     MD

+++ src/scripts/delete_temp_files.sh
@@ -14,20 +14,14 @@ if [ ! -f $orig ]; then
    exit 0
 fi
 
-# we move it away - to prevent conflicts
-temp=${orig}.tmp
-mv $orig $temp
-
-# the last entry is the open_remote_spawn_process calling us - do not delete it!
-last_entry=`tail -1 $temp`
-echo $last_entry >> $orig
-
 # delete the files
-for file in `cat $temp | grep -v -- "$last_entry"`; do
-   rm -f $file
+for file in `cat $orig`; do
+   if [ -f "$file" ]; then
+      rm -f $file
+   fi
 done
 
 # cleanup
-rm $temp
+rm $orig
 
 exit 0
